+++ tests/assembly/test_retain_autoreleased/lib.rs
@@ -5,7 +5,7 @@ use objc2::runtime::{Object, Sel};
 use objc2::MessageReceiver;
 
 #[no_mangle]
-pub fn handle(obj: &Object, sel: Sel) -> Option<Id<Object, Shared>> {
-    let ptr: *mut Object = unsafe { MessageReceiver::send_message(obj, sel, ()).unwrap() };
-    unsafe { Id::retain_autoreleased(ptr) }
+unsafe fn handle(obj: &Object, sel: Sel) -> Option<Id<Object, Shared>> {
+    let ptr: *mut Object = MessageReceiver::send_message(obj, sel, ()).unwrap();
+    Id::retain_autoreleased(ptr)
 }